+++ src/index.css
@@ -6,6 +6,8 @@
 
 
 @theme inline {
+    --container-max-width: 1280px;
+      --container-padding: 1rem;
   --radius-sm: calc(var(--radius) - 4px);
   --radius-md: calc(var(--radius) - 2px);
   --radius-lg: var(--radius);
@@ -119,4 +121,12 @@
   body {
     @apply bg-background text-foreground;
   }
+}
+
+.container {
+  max-width: var(--container-max-width);
+  padding-left: var(--container-padding);
+  padding-right: var(--container-padding);
+  margin-left: auto;
+  margin-right: auto;
 }
