+++ src/layout/RouterLayout.jsx
@@ -1,0 +1,7 @@
+import { Outlet } from 'react-router-dom'
+
+export default function RouterLayout() {
+    return <>
+        <Outlet/>
+    </>
+}
