+++ src/main.jsx
@@ -2,18 +2,12 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.jsx'
-import { Provider } from 'react-redux'
-import { store } from './Redux/store.js'
-import ThemeProvider from './Theme/ThemeProvider.jsx'
+
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <Provider store={store}>
-      <ThemeProvider>
 
         <App />
-      </ThemeProvider>
 
-    </Provider>
   </StrictMode>,
 )
